% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bot.R
\name{Bot}
\alias{Bot}
\title{A Bot Function}
\usage{
Bot(myname = "")
}
\arguments{
\item{myname}{string}
}
\value{
message
}
\description{
This function replies you with a response.
}
\examples{
"hello Thuan! How may I help you?"
Bot()
}
\keyword{chatbot}

