\name{bot_stats}
\alias{bot_stats}
\title{# It compute and display statistique about a campain.}
\description{This function browses bot_cache_dir directory and collects information about tasks. Next, it display gantt chart, tasks chart and how computing element are used. Finally it prints on outpout some stats about the campain.}
\usage{bot_stats(bot_cache_dir = "cache", WHITH4NO = FALSE)}
\arguments{
  \item{bot_cache_dir}{the directory where task results are cached}
  \item{WHITH4NO}{TRUE if you want to include redundante submission in the stats.}
}



\author{Florent Chuffart}





